+++ packages/server/src/api/controllers/Sales/PaymentReceives.ts
@@ -142,7 +142,7 @@ export default class PaymentReceivesController extends BaseController {
       check('entries.*.id').optional({ nullable: true }).isNumeric().toInt(),
       check('entries.*.index').optional().isNumeric().toInt(),
       check('entries.*.invoice_id').exists().isNumeric().toInt(),
-      check('entries.*.payment_amount').exists().isNumeric().toInt(),
+      check('entries.*.payment_amount').exists().isNumeric().toFloat(),
     ];
   }
